Nginx_Autodiscover/via_outlook_host/requirements.txt
CycLog        == 1.1.4
ldap3         == 2.9.1
python-dotenv == 1.0.1
redis         == 5.0.8